requirements.txt
ollama>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
prompt-toolkit>=3.0.0
pyyaml>=6.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6


